+++ includes/zshrc-files/zshrc-linux.zsh
@@ -1,4 +1,4 @@
-####[ Oh-My-Zsh Configurations ]########################################################
+####[ Oh-My-Zsh Configurations ]############################################################
 
 
 # Path to your oh-my-zsh installation.
@@ -43,9 +43,9 @@ HIST_STAMPS="yyyy-mm-dd"
 plugins=(colored-man-pages command-not-found)  # Server
 
 
-####[ Pre `compinit` ]##################################################################
-#### These are configurations that have to be set before the `compinit` function is
-#### called, which is done when sourcing the 'oh-my-zsh.sh' file.
+####[ Pre `compinit` ]######################################################################
+#### These are configurations that have to be set before the `compinit` function is called,
+#### which is done when sourcing the 'oh-my-zsh.sh' file.
 
 
 ## Zsh plugin for completions.
@@ -58,13 +58,105 @@ zsh_completion="${ZSH_CUSTOM:-${ZSH:-~/.oh-my-zsh}/custom}/plugins/zsh-completio
 [[ -f ~/.zfunc/_rustup ]] && fpath+=~/.zfunc
 
 
-####[ Source Oh-My-Zsh ]################################################################
+####[ Resource File - Functions ]###########################################################
+
+
+####
+# Perform the update, upgrade, and cleanup of packages managed by the Apt Package Manager.
+apt_update_and_cleanup() {
+    echo "[INFO] Executing: $ sudo apt update"
+    sudo apt update
+    echo "[INFO] Executing: $ sudo apt upgrade -y"
+    sudo apt upgrade -y
+    echo "[INFO] Executing: $ sudo apt autoremove -y"
+    sudo apt autoremove -y
+    echo "[INFO] Executing: $ sudo apt autoclean"
+    sudo apt autoclean
+}
+
+####
+# Perform the update, upgrade, and cleanup of packages managed by Pacman. Additionally, GRUB
+# configurations are updated to include snapshot changes.
+pacman_update_and_cleanup() {
+    echo "[INFO] Executing: $ sudo pacman -Syu"
+    sudo pacman -Syu
+    echo "[INFO] Executing: $ yay && yay -Yc"
+    yay && yay -Yc
+    echo "[INFO] Executing: $ sudo pkgfile -u"
+    sudo pkgfile -u
+    echo "[INFO] Executing: $ UpdateGrubConfigs"
+    sudo grub-mkconfig -o /boot/grub/grub.cfg
+}
+
+###
+### [ Alias Related Functions ]
+###
+
+####
+# List of Aliases
+alias_lt() {
+    cat <<EOF
+####[ Installed Commands ]######################################################
+
+bandwhich  - Terminal bandwidth utilization tool.
+bat        - A cat(1) clone with wings.
+cheat      - Allows you to create and view interactive cheatsheets on the
+             command-line.
+codespell  - Check code for common misspellings.
+duf        - Disk Usage/Free Utility - a better 'df' alternative.
+fzf        - A command-line fuzzy finder.
+ncdu       - ncdu (NCurses Disk Usage) is a curses-based version of the
+             well-known 'du'.
+pstree     - List processes as a tree.
+tmux       - Terminal multiplexer.
+
+###
+### [ Grouped Commands ]
+###
+
+lt_conversion - List of programs used for converting the formats of videos,
+                images, etc.
+lt_git        - List of programs used for git related commands.
+
+
+####[ Keyboard Combinations ]###################################################
+
+Ctrl + O - Allows you to copy what you are currently typing, via 'Ctrl' + 'O'.
+EOF
+}
+
+####
+# Image and Video Formatters
+alias_lt_conversion() {
+    cat <<EOF
+####[ Image and Video Formatters ]##############################################
+
+ffmpeg - FFmpeg is a collection of libraries and tools to process multimedia
+         content.
+magick - Convert between image formats as well as resize an image, blur, crop,
+         despeckle, dither, draw on, flip, join, re-sample, and much more.
+EOF
+}
+
+####
+# Git Related Commands
+alias_lt_git() {
+    cat <<EOF
+####[ Git Related Commands ]####################################################
+
+lazygit  - Simple terminal UI for git commands.
+git open - Opens the GitHub page for a repo/branch in your browser.
+EOF
+}
+
+
+####[ Source Oh-My-Zsh ]####################################################################
 
 
 source "$ZSH/oh-my-zsh.sh"
 
 
-####[ Aliases ]#########################################################################
+####[ Aliases ]#############################################################################
 
 
 ###
@@ -73,71 +165,42 @@ source "$ZSH/oh-my-zsh.sh"
 
 ## General aliases.
 alias zls="eza"
-alias formatc="find . -name '*.cs' -type f -exec clang-format --style='file:$HOME/Programs/Mine/Formatter Configs/CSharp_clang-format/_clang-format' -i {} +"
-alias update-grub-config="sudo grub-mkconfig -o /boot/grub/grub.cfg"
-alias deletelocalbranches="git branch | grep -v 'main' | xargs git branch -D"
 hash xdg-open 2>/dev/null && alias open="xdg-open"
 
+## Configuration related aliases.
+alias update_grub_config="sudo grub-mkconfig -o /boot/grub/grub.cfg"
+
+## File action related aliases.
+alias format_csharp_code="find . -name '*.cs' -type f -exec clang-format --style='file:$HOME/Programs/Mine/Formatter Configs/CSharp_clang-format/_clang-format' -i {} +"
+alias delete_local_git_branches="git branch | grep -v 'main' | xargs git branch -D"
+
 ## Update based aliases.
-alias updateapt="sudo apt update && sudo apt upgrade -y && sudo apt autoremove -y && sudo apt autoclean"
-alias updatepacman="sudo pacman -Syu && yay && yay -Yc && sudo pkgfile -u && update-grub-config"
+alias update_apt="apt_update_and_cleanup"
+alias update_pacman="pacman_update_and_cleanup"
 
 ## Systemd aliases.
-alias start-bluetooth="sudo systemctl start bluetooth.service"
-alias stop-bluetooth="sudo systemctl stop bluetooth.service"
-alias start-firewalld="sudo systemctl start firewalld.service"
-alias stop-firewalld="sudo systemctl stop firewalld.service"
-alias start-docker="sudo systemctl start docker.service containerd.service"
-alias stop-docker="sudo systemctl stop docker.socket docker.service containerd.service"
+alias start_bluetooth="sudo systemctl start bluetooth.service"
+alias stop_bluetooth="sudo systemctl stop bluetooth.service"
+alias start_firewalld="sudo systemctl start firewalld.service"
+alias stop_firewalld="sudo systemctl stop firewalld.service"
+alias start_docker="sudo systemctl start docker.service containerd.service"
+alias stop_docker="sudo systemctl stop docker.socket docker.service containerd.service"
 
 ###
 ### [ Group 2 ]
 ###
-### Due to the number of commands that I find to be useful, I've created aliases
-### containing some of these commands. They are specifically commands that I don't
-### often use, but are useful to have on hand. Having these aliases allows me to see
-### a list of these commands, without having to commit them to memory.
+### Due to the number of commands that I find to be useful, I've created aliases containing
+### some of these commands. They are specifically commands that I don't often use, but are
+### useful to have on hand. Having these aliases allows me to see a list of these commands,
+### without having to commit them to memory.
 ###
 
-alias lt="echo -e \"
-####[ Installed Commands ]##############################################################
-
-bandwhich  - Terminal bandwidth utilization tool.
-bat        - A cat(1) clone with wings.
-cheat      - Allows you to create and view interactive cheatsheets on the command-line.
-codespell  - Check code for common misspellings.
-duf        - Disk Usage/Free Utility - a better 'df' alternative.
-fzf        - A command-line fuzzy finder.
-ncdu       - ncdu (NCurses Disk Usage) is a curses-based version of the well-known 'du'.
-pstree     - List processes as a tree.
-tmux       - Terminal multiplexer.
-
-####[[ Grouped Commands ]]##############################################################
-
-lt_conversion - List of programs used for converting the formats of videos, images, etc.
-lt_git        - List of programs used for git related commands.
+alias lt="alias_lt"
+alias lt_conversion="alias_lt_conversion"
+alias lt_git="alias_lt_git"
 
 
-####[ Keyboard Combinations ]###########################################################
-
-Ctrl + O - Allows you to copy what you are currently typing, via 'Ctrl' + 'O'.
-\""
-alias lt_conversion="echo -e \"
-####[ Image and Video Formatters ]######################################################
-
-ffmpeg - FFmpeg is a collection of libraries and tools to process multimedia content.
-magick - Convert between image formats as well as resize an image, blur, crop,
-         despeckle, dither, draw on, flip, join, re-sample, and much more.
-\""
-alias lt_git="echo -e \"
-####[ Git Related Commands ]############################################################
-
-lazygit  - Simple terminal UI for git commands.
-git open - Opens the GitHub page for a repo/branch in your browser.
-\""
-
-
-####[ Environmental Variables ]#########################################################
+####[ Environmental Variables ]#############################################################
 
 
 # 1Password auth socket.
@@ -166,7 +229,7 @@ fi
 export NVM_DIR="$HOME/.nvm"
 
 
-####[ Sourced Files ]###################################################################
+####[ Sourced Files ]#######################################################################
 
 
 ## Load NVM.
@@ -186,18 +249,19 @@ zsh_autosuggestions="${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggesti
 ## Zsh plugin for fzf-tab.
 ## This plugin is installed via chezmoi, specified in the '.chezmoiexternal.toml' file.
 fzf_tab="${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/fzf-tab/fzf-tab.plugin.zsh"
-[[ -f $fzf_tab ]] && hash fzf 2>/dev/null && source "$fzf_tab"
+[[ -f $fzf_tab ]] \
+    && hash fzf 2>/dev/null \
+    && source "$fzf_tab"
 
 ## Source the plugins.sh file for the `op` command.
-[[ -f $HOME/.config/op/plugins.sh ]] \
-    && source "$HOME/.config/op/plugins.sh"
+[[ -f $HOME/.config/op/plugins.sh ]] && source "$HOME/.config/op/plugins.sh"
 
 ## Make the 'command-not-found' functionality available.
 [[ -f /usr/share/doc/pkgfile/command-not-found.zsh ]] \
     && source /usr/share/doc/pkgfile/command-not-found.zsh
 
 
-####[ Zsh Style Configurations ]########################################################
+####[ Zsh Style Configurations ]############################################################
 
 
 # Disable sort when completing `git checkout`.
@@ -214,9 +278,9 @@ hash eza 2>/dev/null \
 zstyle ':fzf-tab:*' switch-group '<' '>'
 
 
-####[ End of File Configurations ]######################################################
-#### These are configurations that are specified to be placed at the end of the file, by
-#### the developer/documentation.
+####[ End of File Configurations ]##########################################################
+#### These are configurations that are specified to be placed at the end of the file, by the
+#### developer/documentation.
 
 
 # Initialize Starship prompt, if it is installed and $ZSH_THEME is not set.
@@ -225,6 +289,5 @@ hash starship 2>/dev/null \
     && eval "$(starship init zsh)"
 
 
-####[ Others ]##########################################################################
+####[ Others ]##############################################################################
 #### These are generally configurations set up by setup scripts or other programs.
-